requirements.txt
google-cloud-bigquery
functions-framework==3.*
Flask==2.0.2
